refactor: major refactor to use single input component
MAJOR CHANGE (avoiding saying breaking because commitizen will bump the major version and we're not ready for v1 yet, but this WILL BREAK any previous implementations).
All inputs are now accessed via a single component: `<zoa-input>`. A type is specified (if none, it's used as an "empty" spacing/grid component) and options can be passed.
- allows forms to be defined dynamically with a json schema
- makes it much easier to change labels etc because it's all defined in a single place
Buttons are still separate components, not accessed via zoa-input.