fix: also debounce model value change
the "change" event on ZoaInput (i.e. the one the user is probably listening to) is triggered by "update:modelValue" from the nested input component. This change event was being debounced by 0ms because there's no delay on the wrapper, and the update event wasn't being debounced at all, so there was effectively no delay/debounce on the whole chain.

Closes: #36