fix: separate displayed search term and actual search term
The value displayed to the user needs to be updated immediately upon typing, but the search function shouldn't be triggered too often. There are two ways to run a search: the internal search in `dropdownOptions`, or something external and user-defined that runs on the `@search` event. Debouncing the update of the search term (i.e. `activeSearch`) is only useful for the internal search.