fix: emit custom events from input wrapper
Search events were not available on the input wrapper, only on the child component itself. This makes them available on the parent/wrapper by pairing custom events with a generic "zoaEvent" event, which then gets captured by the wrapper and re-emitted under the custom name.
The custom event (i.e. `emit('search')`) has also been kept so that the event can still be listened for if the child component is used without the input wrapper.