refactor(input-parts): fix linter errors, remove injects
Props didn't need to be assigned to a variable. Injects have been removed because they seemed unnecessary and inconsistent; these two elements are always directly inside a ZoaInput component so we don't need to avoid prop-drilling, and we were injecting some and passing others in as regular props so it was strangely inconsistent.