fix(toggle-button): simplify toggle button
The toggle button was previously a checkbox made to look like a button, which was overcomplicated and made the purpose unclear (especially for screenreaders). It's now simply an a element styled to look like a button and can only output a boolean value, not an array or a string. checkValue and name have been removed and you can no longer use this in a group like a checkbox.

BREAKING CHANGE: simplify toggle button